Factor out common functionality
