+++ src/gz_dashboard/commands/sync.py
@@ -2,16 +2,15 @@ import argparse
 import copy
 import os
 import sys
-import urllib.request as request
 
-import yaml
 from humanfriendly.terminal import ansi_wrap
 from vcstool.commands.import_ import add_dependencies, generate_jobs, get_repositories
 from vcstool.executor import execute_jobs, output_results
 
-from gz_dashboard import __version__ as dashboard_version
 from gz_dashboard.commands.command import add_common_arguments
+from gz_dashboard.distribution import get_distributions
 from gz_dashboard.streams import set_streams
+from gz_dashboard.util import file_or_url_type, resolve_uri
 
 
 def get_parser():
@@ -45,41 +44,8 @@ def existing_dir(path):
     return path
 
 
-def file_or_url_type(value):
-    print(value)
-    if os.path.exists(value) or "://" not in value:
-        return argparse.FileType("r")(value)
-    # use another user agent to avoid getting a 403 (forbidden) error,
-    # since some websites blacklist or block unrecognized user agents
-    return request.Request(
-        value, headers={"User-Agent": f"gz-dashboard/{dashboard_version}"}
-    )
-
-
-def get_distributions(yaml_file):
-    try:
-        root = yaml.safe_load(yaml_file)
-    except yaml.YAMLError as e:
-        raise RuntimeError(f"Input data is not valid yaml format: {e}")
-
-    try:
-        distributions = root["distributions"]
-        return distributions
-    except KeyError as e:
-        raise RuntimeError(f"Input data is not valid format: {e}")
-
-
 def sync_distro(name, config, args):
-    print(f"sync_distro: {name}, {config}")
-    config = file_or_url_type(config["url"])
-    try:
-        if isinstance(config, request.Request):
-            config = request.urlopen(config)
-    except (RuntimeError, request.URLError) as e:
-        print(ansi_wrap(str(e), color="red"), file=sys.stderr)
-        return 1
-
-    print(config)
+    config = resolve_uri(config["url"])
 
     args.path = os.path.join(args.path, name)
     args.recursive = False
@@ -100,10 +66,8 @@ def sync_distro(name, config, args):
     results = execute_jobs(
         jobs, show_progress=True, number_of_workers=args.workers, debug_jobs=args.debug
     )
-
     output_results(results)
     any_error = any(r["returncode"] for r in results)
-
     return 1 if any_error else 0
 
 
@@ -116,16 +80,7 @@ def main(args=None, stdout=None, stderr=None):
     if args is None:
         args = sys.argv[1:]
     args = parser.parse_args(args)
-
-    try:
-        config = args.config
-        if isinstance(config, request.Request):
-            config = request.urlopen(config)
-    except (RuntimeError, request.URLError) as e:
-        print(ansi_wrap(str(e), color="red"), file=sys.stderr)
-        return 1
-
-    distros = get_distributions(config)
+    distros = get_distributions(args.config)
 
     for (distro_name, distro_config) in distros.items():
         res = sync_distro(distro_name, distro_config, copy.copy(args))
